move shared code to common
