+++ source/extensions/credentials/basic_auth/BUILD
@@ -17,6 +17,7 @@ envoy_cc_library(
         "//source/common/common:base64_lib",
         "//source/common/config:datasource_lib",
         "//source/extensions/credentials/common:credential_interface",
+        "//source/extensions/credentials/common:secret_reader_lib",
     ],
 )
 
@@ -28,6 +29,7 @@ envoy_cc_extension(
         ":basic_auth_lib",
         "//source/common/http:headers_lib",
         "//source/extensions/credentials/common:credential_factory_base_lib",
+        "//source/extensions/credentials/common:secret_reader_lib",
         "@envoy_api//envoy/extensions/credentials/basic_auth/v3:pkg_cc_proto",
     ],
 )
