+++ source/extensions/credentials/common/BUILD
@@ -17,6 +17,14 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
+    name = "secret_reader_lib",
+    hdrs = ["secret_reader.h"],
+    deps = [
+        "//source/common/config:datasource_lib",
+    ],
+)
+
+envoy_cc_library(
     name = "credential_config_interface",
     hdrs = ["factory.h"],
     deps = [
